+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/RelluEssentials.java
@@ -4,7 +4,7 @@ import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.Multimap;
 import de.relluem94.minecraft.server.spigot.essentials.api.*;
 import de.relluem94.minecraft.server.spigot.essentials.commands.*;
-import de.relluem94.minecraft.server.spigot.essentials.constants.MessageKey;
+import de.relluem94.minecraft.server.spigot.essentials.enums.MessageKey;
 import de.relluem94.minecraft.server.spigot.essentials.events.*;
 import de.relluem94.minecraft.server.spigot.essentials.events.bag.BlockBreakBags;
 import de.relluem94.minecraft.server.spigot.essentials.events.bag.BlockDropItemBags;
@@ -14,9 +14,9 @@ import de.relluem94.minecraft.server.spigot.essentials.events.npc.*;
 import de.relluem94.minecraft.server.spigot.essentials.events.protect.*;
 import de.relluem94.minecraft.server.spigot.essentials.helpers.DatabaseHelper;
 import de.relluem94.minecraft.server.spigot.essentials.helpers.LanguageHelper;
-import de.relluem94.minecraft.server.spigot.essentials.helpers.objects.Selection;
-import de.relluem94.minecraft.server.spigot.essentials.helpers.pojo.*;
 import de.relluem94.minecraft.server.spigot.essentials.managers.*;
+import de.relluem94.minecraft.server.spigot.essentials.model.Selection;
+import de.relluem94.minecraft.server.spigot.essentials.model.pojo.*;
 import de.relluem94.minecraft.server.spigot.essentials.npc.*;
 import de.relluem94.minecraft.server.spigot.essentials.npc.trader.Banker;
 import de.relluem94.minecraft.server.spigot.essentials.wrapper.CommandWrapper;
