+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/enums/PlayerState.java
@@ -1,4 +1,4 @@
-package de.relluem94.minecraft.server.spigot.essentials.constants;
+package de.relluem94.minecraft.server.spigot.essentials.enums;
 
 import de.relluem94.minecraft.server.spigot.essentials.interfaces.IPlayerState;
